+++ src/ChaiLove.cpp
@@ -7,6 +7,7 @@
 ChaiLove* ChaiLove::m_instance = NULL;
 retro_input_state_t ChaiLove::input_state_cb = NULL;
 retro_input_poll_t ChaiLove::input_poll_cb = NULL;
+retro_environment_t ChaiLove::environ_cb = NULL;
 
 void ChaiLove::destroy() {
 	m_instance = NULL;
@@ -58,6 +59,7 @@ bool ChaiLove::load(const std::string& file) {
 		sound.load();
 	}
 
+	console.load(config);
 	graphics.load();
 	image.load();
 	keyboard.load();
@@ -131,6 +133,9 @@ void ChaiLove::draw() {
 		#ifdef __HAVE_TESTS__
 		test.draw();
 		#endif
+
+		// Render the in-game console.
+		console.draw();
 
 		// Flip the buffer.
 		if (SDL_Flip(screen) == -1) {