+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to [ChaiLove](https://github.com/RobLoach/ChaiLove) will be 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.17.0 - Unreleased
+### Added
+- `love.system.getOS()` now returns actual operating system name
+- `love.console` an in-game console toggled with the tilde key
+  - Enable the console with `love.console.setEnabled(true)`
+- `love.graphics.getColor()` to retrieve the active painting color
+- `love.window.showMessageBox()` to display on-screen messages
+
+### Fixed
+- Cleaned the Makefile dependency chain
+
 ## 0.16.1 - 2018-03-12
 ### Added
 - `List()` for linked list